+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## Unreleased
 
 - Improved `GetAccount` performance if number of vault assets exceeds the limit. ([#2227](https://github.com/0xMiden/node/pull/2227)).
+- Improved `GetAccount` performance if number of storage map entries exceeds the limit. ([#2228](https://github.com/0xMiden/node/pull/2228)).
+
 
 ## v0.15.0-rc.4 (2026-06-09)